Add support for reference types proposal
This adds support for the reference type proposal. This adds support for
all reference types (anyref, funcref(=anyfunc), and nullref) and four
new instructions: `ref.null`, `ref.is_null`, `ref.func`, and new typed
`select`. This also adds subtype relationship support between reference
types.

This does not include table instructions yet. This also does not include
wasm2js support.

This does not pass `fuzz_opt.py` yet because there are remaining bugs
related to `nullref` handling, because `nullref` is not allowed to be
written to text or binary format. We are discussing whether we should
allow it to be written in WebAssembly/reference-types#60 now.

This is not ready for review yet because it does not pass fuzzer now;
I'm uploading this just to show how the current a bit ad-hoc `nullref`
handling looks like. So don't spend too much time on reviewing each
detail for now. I apprciate some high-level comments on how we should
handle `nullref` or other stuff.